// app.client/src/api/models/UpdateContractCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateContractCommand = {
    contractId?: string | null;
    name?: string | null;
    description?: string | null;
    startDate?: string | null;
    endDate?: string | null;
    number?: string | null;
    companyID?: string | null;
};